# Remove q concat in FA3 backend for DeepSeek decode