<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

concat for q in FA3 backend can be removed since the interface for q_nope and q_rope are separate.

### Accuracy
```
python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-remote-code --attention-backend fa3 --disable-radix
python3 benchmark/gsm8k/bench_sglang.py --num-questions 1400 --parallel 160 --num-shots 8
```
main:
```
Accuracy: 0.944
Invalid: 0.000
Latency: 252.757 s
Output throughput: 567.307 token/s
```
this PR:
```
Accuracy: 0.947
Invalid: 0.000
Latency: 246.356 s
Output throughput: 583.927 token/s
```
### Performance
```
python3 -m sglang.launch_server --model /dev/shm/DeepSeek-V3 --tp 8 --trust-remote-code --attention-backend fa3 --disable-radix
python3 -m sglang.bench_one_batch_server --model None --base-url http://0.0.0.0:30000 --batch-size 1 --input-len 512 --output-len 512
```
main:
```
batch size: 1
latency: 10.38 s
output throughput: 49.30 token/s
(input + output) throughput: 98.61 token/s

batch size: 32
latency: 18.97 s
output throughput: 863.68 token/s
(input + output) throughput: 1727.35 token/s
```
this PR:
```
batch size: 1
latency: 10.29 s
output throughput: 49.75 token/s
(input + output) throughput: 99.51 token/s

batch size: 32
latency: 18.15 s
output throughput: 902.83 token/s
(input + output) throughput: 1805.66 token/s
```
1% improvement for bs 1 and 4% for bs 32.
### Profile
main:
<img width="564" alt="image" src="https://github.com/user-attachments/assets/0ec0cc52-9c31-4baf-a739-c670aacd33d1" />

this PR:
<img width="575" alt="image" src="https://github.com/user-attachments/assets/b311f5ca-2ccf-47e3-9deb-14579a2dcea4" />

2.5μm reduced for bs 1.